refactor(cycle54): extract shared sanitize.cjs (fixes latent bug + a11y)
Cycle 54 (autonomous loop; directive "продолжай").

## Defect (real, measured)
`scripts/sanitizeExternal` was DUPLICATED in both sync-github.cjs and
sanitize-generated.cjs — and BOTH copies carried a LATENT bug: the
javascript:/vbscript: link-scheme regex
  /(\[[^\]]*\]\(\)\s*(javascript:|vbscript:)[^)\s]*/gi
had an ESCAPED `\)` so group 1 was never closed. It would have thrown
"Unterminated group" the moment a `javascript:` markdown link was synced
(dormant only because sync rarely hits one). Also neither copy padded
missing `<img alt>`, so generated gist pages tripped the (non-blocking)
a11y spot-check (4 warnings).

## Fix
- Extracted a single source of truth: `scripts/sanitize.cjs` (CommonJS .cjs,
  because the repo is `"type": "module"` and `module.exports` needs .cjs).
  It strips `<script>` / inline `on*=` / `javascript:`+`vbscript:` link
  schemes / `<iframe>`+`<object>`+`<embed>`, AND pads missing `<img> alt`.
- Replaced both duplicated inline functions with `require('./sanitize.cjs')`.
- Verified: `javascript:`/`vbscript:` -> `(#)`, `<script>` stripped,
  `<img>` gets `alt`, legitimate code blocks (`a < b && c > d`) preserved.
- Ran sanitize-generated.cjs over existing generated content: the 4 `<img>`
  alt a11y warnings on the gist page dropped to 0.

## Verification
lint clean · jest 277/21 · hugo build 0 · check-html 0 warnings.
Reverted the generated gist-file edits (regenerated by CI with the fix).
