+++ .beads/beads.json
@@ -1364,6 +1364,13 @@
       "status": "done",
       "pr": null,
       "note": "Discovered scripts/audit-internal-links.cjs (source-level markdown internal-link auditor) was NEVER invoked by any workflow, package.json script, or other script -- it was dead code. Its cycle-43 IGNORE list (8 mirrored awesome-lists) and 0-BROKEN result were therefore unused in CI. Wired it into .github/workflows/quality.yml as a report-only step after the built-surface check-links.cjs, giving pre-build source-level link feedback (complements Hugo alias/redirect resolution). Verified: script exits 0, 0 actionable dead links on current content. Non-blocking (no process.exit(1)), so pipeline safe. This is a workflow edit -> Quality CI needs manual dispatch on the PR."
+    },
+    {
+      "id": "T111",
+      "title": "cycle54: extract shared sanitize.cjs (fixes latent javascript: regex bug + adds img alt)",
+      "status": "done",
+      "pr": null,
+      "note": "Discovered scripts/sanitizeExternal was DUPLICATED in sync-github.cjs and sanitize-generated.cjs, AND both copies had a LATENT bug: the javascript:/vbscript: link-scheme regex /(\\[[^\\]]*\\]\\(\\)\\s*(javascript:|vbscript:)[^)\\s]*/gi had an escaped \\) so group 1 was never closed -> would throw if it ever hit a javascript: markdown link (dormant because sync rarely does). Extracted a single source of truth scripts/sanitize.cjs (CommonJS .cjs so module.exports works under the repo's type:module). It strips <script>/on*=/javascript:/vbscript:/iframe+ pads missing <img> alt. Wired into both importers; removed the duplicated inline functions. Verified: javascript:/vbscript: -> (#), <script> stripped, <img> gets alt, code blocks preserved. Also ran sanitize-generated.cjs over existing generated content -> 4 img-alt a11y warnings on the gist page dropped to 0. Reverted the generated gist-file edits (regenerated by CI with the fix). Full gate: lint clean, jest 277/21, build 0, check-html 0 warnings."
     }
   ],
   "edges": [
@@ -1660,6 +1667,11 @@
     {
       "from": "T109",
       "to": "T110",
+      "type": "next"
+    },
+    {
+      "from": "T110",
+      "to": "T111",
       "type": "next"
     }
   ]
